Changing the manger view.
This commit updates the manger view by changing the location of the buttons to the match the order of the use cases. I also add some styling to the buttons
so all the buttons are the same size. This commit also changes and spelling error the modifyInventoryItemTypeBTN. The y in the button name was a t isntead
of an t. So I changed the spelling in all the locations to match the new spelling.