Creating vendor inventory item type model and view.
WIP: This commit creates the necessary view and model for the vendor inventory item type transaction. The inserting
into the database still needs to be tested. I was having trouble getting to the view I created so some work still
needs to be done in order to make sure the vendor inventory item type is getting inserted into the database.