Implementing choice selection.
In order to insert a vendory inventory item type the user needs to select a vendor and an inventory item type.
There wasn't an easy solution to pass this information around so I made static variables on the classes in order
to keep this information untill I need it on the vendor price view screen. This solution can be changed if we
come up with a better way to pass the primary keys of the selected options.