Implementing saving of the selector vendor.
This commit uses the methods introduced in the vendor class to save the selected vendor. It also leaves a
note about the state change request not calling the manger state change request. I am not sure why this is
so I left a commment about it as it needs to be addressed.