Updating manager model and view factory.
This commit upates the manager model and the view factory for the new vendor inventory item type model and view.
This commit updates how the Manger.choice is set. Before, it was being set on every state change request, but this
is wrong. The key should only be set if the state change request is a top level request. One that is from the choice
transaction screen because the choice shouldn't be updated when we switch to vendor collection only the orginal choice
should be saved.I also implemented history in the vender selected key to go to the correct screen based on what the
orginally choice was. I haven't been able to test this because the vendor collection view is not calling the manager
state change request for some reason. This still needs to be tested.