Working on Add InventoryItemType.
This commit is working on the Add InventoryItemType usecase. I fixed an error in the
AddIITView class and added the components to the grid. I also started working on
inserting it into the database, but the model needs to be updated. I also changed the
manager and the view factory to wire up the view.